// src/app/registerPatient/patient.type.ts
export interface typeSubmit {
    name: string
    email: string
    sexo: "M" | "F"
    cpf: string
    rg: string
    birth: string
    type_blood: string
    agreement: string
    medical_agreement_number: string
    phone: string
    zipcode: string
    street: string
    complement: string
    city: string
    state: string
}